feat: replace database config params with URL
