+++ cmd/migrate/v3/types.go
@@ -1,7 +1,6 @@
 package v3
 
 import (
-	databaseconfig "github.com/forbole/juno/v3/database/config"
 	loggingconfig "github.com/forbole/juno/v3/logging/config"
 	"github.com/forbole/juno/v3/modules/pruning"
 	"github.com/forbole/juno/v3/modules/telemetry"
@@ -11,17 +10,30 @@ import (
 	"github.com/forbole/juno/v3/types/config"
 )
 
-// Config defines all necessary juno configuration parameters.
 type Config struct {
-	Chain    config.ChainConfig    `yaml:"chain"`
-	Node     nodeconfig.Config     `yaml:"node"`
-	Parser   parserconfig.Config   `yaml:"parsing"`
-	Database databaseconfig.Config `yaml:"database"`
-	Logging  loggingconfig.Config  `yaml:"logging"`
+	Chain    config.ChainConfig   `yaml:"chain"`
+	Node     nodeconfig.Config    `yaml:"node"`
+	Parser   parserconfig.Config  `yaml:"parsing"`
+	Database DatabaseConfig       `yaml:"database"`
+	Logging  loggingconfig.Config `yaml:"logging"`
 
 	// The following are there to support modules which config are present if they are enabled
 
 	Telemetry *telemetry.Config       `yaml:"telemetry,omitempty"`
 	Pruning   *pruning.Config         `yaml:"pruning,omitempty"`
 	PriceFeed *pricefeedconfig.Config `yaml:"pricefeed,omitempty"`
+}
+
+type DatabaseConfig struct {
+	Name               string `yaml:"name"`
+	Host               string `yaml:"host"`
+	Port               int64  `yaml:"port"`
+	User               string `yaml:"user"`
+	Password           string `yaml:"password"`
+	SSLMode            string `yaml:"ssl_mode,omitempty"`
+	Schema             string `yaml:"schema,omitempty"`
+	MaxOpenConnections int    `yaml:"max_open_connections"`
+	MaxIdleConnections int    `yaml:"max_idle_connections"`
+	PartitionSize      int64  `yaml:"partition_size,omitempty"`
+	PartitionBatchSize int64  `yaml:"partition_batch,omitempty"`
 }
