+++ cmd/migrate/v3/utils.go
@@ -1,4 +1,4 @@
-package v1
+package v3
 
 import (
 	"fmt"
@@ -6,14 +6,14 @@ import (
 	"os"
 	"path"
 
-	"github.com/pelletier/go-toml"
+	"gopkg.in/yaml.v3"
 
 	"github.com/forbole/juno/v3/types/config"
 )
 
-// GetConfig returns the configuration reading it from the config.toml file present inside the home directory
+// GetConfig returns the configuration reading it from the config.yaml file present inside the home directory
 func GetConfig() (Config, error) {
-	file := path.Join(config.HomePath, "config.toml")
+	file := path.Join(config.HomePath, "config.yaml")
 
 	// Make sure the path exists
 	if _, err := os.Stat(file); os.IsNotExist(err) {
@@ -26,6 +26,6 @@ func GetConfig() (Config, error) {
 	}
 
 	var cfg Config
-	err = toml.Unmarshal(bz, &cfg)
+	err = yaml.Unmarshal(bz, &cfg)
 	return cfg, err
 }
