+++ database/config/config.go
@@ -1,33 +1,30 @@
 package config
 
+import "net/url"
+
 type Config struct {
-	Name               string `yaml:"name"`
-	Host               string `yaml:"host"`
-	Port               int64  `yaml:"port"`
-	User               string `yaml:"user"`
-	Password           string `yaml:"password"`
-	SSLMode            string `yaml:"ssl_mode,omitempty"`
-	Schema             string `yaml:"schema,omitempty"`
+	URL                string `yaml:"url"`
 	MaxOpenConnections int    `yaml:"max_open_connections"`
 	MaxIdleConnections int    `yaml:"max_idle_connections"`
 	PartitionSize      int64  `yaml:"partition_size"`
 	PartitionBatchSize int64  `yaml:"partition_batch"`
 }
 
+func (c *Config) GetUser() string {
+	parsedURL, err := url.Parse(c.URL)
+	if err != nil {
+		panic(err)
+	}
+	return parsedURL.User.Username()
+}
+
 func NewDatabaseConfig(
-	name, host string, port int64, user string, password string,
-	sslMode string, schema string,
+	url string,
 	maxOpenConnections int, maxIdleConnections int,
 	partitionSize int64, batchSize int64,
 ) Config {
 	return Config{
-		Name:               name,
-		Host:               host,
-		Port:               port,
-		User:               user,
-		Password:           password,
-		SSLMode:            sslMode,
-		Schema:             schema,
+		URL:                url,
 		MaxOpenConnections: maxOpenConnections,
 		MaxIdleConnections: maxIdleConnections,
 		PartitionSize:      partitionSize,
@@ -38,13 +35,7 @@ func NewDatabaseConfig(
 // DefaultDatabaseConfig returns the default instance of Config
 func DefaultDatabaseConfig() Config {
 	return NewDatabaseConfig(
-		"database-name",
-		"localhost",
-		5432,
-		"user",
-		"password",
-		"",
-		"public",
+		"postgresql://user:password@localhost:5432/database-name?sslmode=disable&search_path=public",
 		1,
 		1,
 		100000,
